+++ src/components/ImageHandlerContext/context.tsx
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 import { ConfigurationContextType } from "../../types";
 
-export const ConfigurationContext = createContext<
-  ConfigurationContextType | {}
->({});
+export const ConfigurationContext = createContext<ConfigurationContextType>({
+  endpointDomain: "",
+});