+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.0] – 2022-11-
 
+- Added global context with configuration options and custom type for configuration context options.
+- Added image component to output the image.
+- Added type for available image filters.
+- Added general helper functions: `checkFiletype()`
+- Added function `mapFilterObjectToUrl()` to map for filters a string for the URL.
+- Added function `generateImgSrc()` to create image source URLs.
+- Added function `prepareSrc()` to prepare the `src` value before using it.
 - Added a browserslist to the `package.json` file.
 
 ## [0.1.0] – 2022-11-24